refactor stream/chat/embed-stram to be a single execution logic path so that it is easier to maintain and build upon
